+++ cloudmanager/data_source_netapp_cloudmanager_volume.go
@@ -171,7 +171,7 @@ func dataSourceCVOVolumeRead(d *schema.ResourceData, meta interface{}) error {
 	client := meta.(*Client)
 	clientID := d.Get("client_id").(string)
 	volume := volumeRequest{}
-	weInfo, err := client.getWorkingEnvironmentDetail(d, clientID)
+	weInfo, err := client.getWorkingEnvironmentDetail(d, clientID, true, "")
 	if err != nil {
 		return fmt.Errorf("cannot find working environment")
 	}
@@ -180,7 +180,7 @@ func dataSourceCVOVolumeRead(d *schema.ResourceData, meta interface{}) error {
 	volume.SvmName = weInfo.SvmName
 	d.Set("working_environment_id", volume.WorkingEnvironmentID)
 	d.Set("svm_name", volume.SvmName)
-	res, err := client.getVolume(volume, clientID)
+	res, err := client.getVolume(volume, clientID, true, "")
 	if err != nil {
 		log.Print("Error reading volume")
 		return err